Lib: Split trait & impls into different files
