+++ faebryk/library/__init__.py
@@ -6,3 +6,4 @@ import faebryk.library.core
 import faebryk.library.kicad
 import faebryk.library.library
 import faebryk.library.traits
+import faebryk.library.trait_impl
