+++ faebryk/library/library/parameters.py
@@ -5,7 +5,7 @@ import logging
 
 logger = logging.getLogger("library")
 
-from faebryk.library.traits.parameter import is_representable_by_single_value
+from faebryk.library.trait_impl.parameter import is_representable_by_single_value_defined
 from faebryk.library.core import Parameter
 from faebryk.library.traits import *
 from faebryk.libs.exceptions import FaebrykException
@@ -16,7 +16,7 @@ class Constant(Parameter):
     def __init__(self, value: typing.Any) -> None:
         super().__init__()
         self.value = value
-        self.add_trait(is_representable_by_single_value(self.value))
+        self.add_trait(is_representable_by_single_value_defined(self.value))
 
 
 class Range(Parameter):
@@ -31,7 +31,7 @@ class Range(Parameter):
                 f"Value not in range: {value_to_check} not in [{self.min},{self.max}]"
             )
 
-        self.add_trait(is_representable_by_single_value(value_to_check))
+        self.add_trait(is_representable_by_single_value_defined(value_to_check))
 
 
 class TBD(Parameter):
