+++ faebryk/library/trait_impl/__init__.py
@@ -1,0 +1,8 @@
+# This file is part of the faebryk project
+# SPDX-License-Identifier: MIT
+
+import faebryk.library.trait_impl.component
+import faebryk.library.trait_impl.footprint
+import faebryk.library.trait_impl.interface
+import faebryk.library.trait_impl.link
+import faebryk.library.trait_impl.parameter
