+++ faebryk/library/trait_impl/footprint.py
@@ -1,0 +1,9 @@
+# This file is part of the faebryk project
+# SPDX-License-Identifier: MIT
+
+import logging
+
+logger = logging.getLogger("library")
+
+from faebryk.library.core import FootprintTrait
+from faebryk.library.traits.footprint import *
