+++ faebryk/library/trait_impl/interface.py
@@ -1,0 +1,8 @@
+# This file is part of the faebryk project
+# SPDX-License-Identifier: MIT
+
+import logging
+
+logger = logging.getLogger("library")
+
+from faebryk.library.traits.interface import *
