+++ faebryk/library/trait_impl/link.py
@@ -1,0 +1,8 @@
+# This file is part of the faebryk project
+# SPDX-License-Identifier: MIT
+
+import logging
+
+logger = logging.getLogger("library")
+
+from faebryk.library.traits.link import *
