+++ faebryk/library/trait_impl/parameter.py
@@ -1,0 +1,18 @@
+# This file is part of the faebryk project
+# SPDX-License-Identifier: MIT
+
+import logging
+
+logger = logging.getLogger("library")
+
+import typing
+from faebryk.library.traits.parameter import *
+
+
+class is_representable_by_single_value_defined(is_representable_by_single_value):
+    def __init__(self, value: typing.Any) -> None:
+        super().__init__()
+        self.value = value
+
+    def get_single_representing_value(self):
+        return self.value
