+++ faebryk/library/traits/parameter.py
@@ -10,9 +10,5 @@ from faebryk.library.core import ParameterTrait
 
 
 class is_representable_by_single_value(ParameterTrait):
-    def __init__(self, value: typing.Any) -> None:
-        super().__init__()
-        self.value = value
-
     def get_single_representing_value(self):
-        return self.value
+        return NotImplementedError
