+++ samples/resistors_and_nand.py
@@ -30,7 +30,7 @@ def run_experiment():
     from faebryk.library.library.footprints import SMDTwoPin
     from faebryk.library.library.interfaces import Power
     from faebryk.library.library.parameters import Constant
-    from faebryk.library.traits.component import (
+    from faebryk.library.trait_impl.component import (
         has_defined_footprint,
         has_symmetric_footprint_pinmap,
     )
